Added support for non-ascii filenames in content-disposition.
Using the details outlined at:
http://kbyanc.blogspot.hk/2010/07/serving-file-downloads-with-non-ascii.html

To support older browsers, an ascii only filename is added to the
content-dispostion. The unicode version is properly encoded and added
to the header. The argument attachment_filename may now be False to
send no filename. Added and fixed unit tests to verify the new
behavior.